+++ backend/modules/template_module/adapters/repositories/sqlalchemy/__init__.py
@@ -1,4 +1,4 @@
-from .repository import (
+from .repositories import (
     SqlAlchemyTemplateDomainRepository,
     SqlAlchemyTemplateQueryRepository,
 )
